APT-293 fix problem with Confirm response

Approved-by: Kamil Kolarik <kamil.kolarik@thinkcreatix.com>

+++ backend/src/main/java/com/creatix/service/MaintenanceReservationService.java
@@ -137,9 +137,11 @@ public class MaintenanceReservationService {
                 }
                 if (reservation.getId() != response.getSlotUnitId()) {
                     releaseReservedCapacity(reservation);
-                    notification.getReservations().remove(reservation);
-                    reservationIterator.remove();
-                    reservationDao.delete(reservation);
+                    if (MaintenanceNotificationResponseRequest.ResponseType.Confirm.equals(response.getResponse())) {
+                        notification.getReservations().remove(reservation);
+                        reservationIterator.remove();
+                        reservationDao.delete(reservation);
+                    }
                 }
             }
         }
